# new-iVan-master/docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swipped API
  version: 1.0.0
  description: |
    ## Introduction
    Unified API documentation for Swipped.

    This document groups endpoints by business domain so API consumers can navigate
    authentication, public discovery, and protected/admin operations more easily.

    ## Base URL
    - Live: `https://dev.swippped.com`
    - Local: `http://localhost:3001`
servers:
  - url: https://dev.swippped.com
    description: Live domain
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/business-account/apply:
    post:
      tags: [Public]
      summary: Create/submit operation
      requestBody: *defaultBody
      responses: *stdResponses
...
  /api/cart:
    get: &customerGet
      tags: [Customer]
      summary: "Customer - Cart: Get cart"
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses: *stdResponses
    post: &customerPost
      tags: [Customer]
      summary: "Customer - Cart: Add or update cart item"
      requestBody: *defaultBody
      responses: *stdResponses
    delete: &defaultDelete
      tags: [Admin]
      summary: Delete operation
      responses: *stdResponses
...
  /api/chats/support:
    post: *customerPost
...
components:
...
  parameters:
...
    SearchParam:
      name: search
      in: query
      required: false
      schema:
        type: string
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1